+++ Core/AbstractServices/IAuthenticationServices.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Win32;
-using Shared.Dtos.Identity;
+using Shared.Dtos.Identity.Login;
+using Shared.Dtos.Identity.Register;
 using System;
 using System.Collections.Generic;
 using System.Linq;
