+++ Shared/Dtos/Identity/Login/LoggedInBase.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Shared.Dtos.Identity
+namespace Shared.Dtos.Identity.Login
 {
-	public abstract class LoggedIn
+	public abstract class LoggedInBase
 	{
 		public string UserName { get; set; } = default!;
 		public string Email { get; set; } = default!;
