+++ Shared/Dtos/Identity/Register/RegisterUserDto.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Shared.Dtos.Identity
+namespace Shared.Dtos.Identity.Register
 {
 	public class RegisterUserDto : RegisterBaseDto
 	{
